+++ packages/core/src/schema/__tests__/converter.test.ts
@@ -999,4 +999,214 @@ describe('convertZodToJsonSchema', () => {
       expect(result.hasUnmappedNodes).toBe(false);
     });
   });
+
+  describe('fallback behavior and x-zod annotations', () => {
+    describe('x-zod.unmapped fallback', () => {
+      it('should emit x-zod.unmapped for unsupported Zod types', () => {
+        // Create a schema with an unsupported type
+        const schema = z.string();
+
+        // Mock the type to be unsupported
+        Object.defineProperty(schema._def, 'typeName', {
+          value: 'ZodUnsupported',
+          writable: false,
+        });
+
+        const result = convertZodToJsonSchema(schema);
+
+        expect(result.schema).toEqual({
+          'x-zod': {
+            unmapped: true,
+            reason: 'Unsupported type: ZodUnsupported',
+          },
+        });
+        expect(result.hasUnmappedNodes).toBe(true);
+        expect(
+          result.warnings.some(w => w.includes('Unsupported Zod type'))
+        ).toBe(true);
+      });
+
+      it('should handle lazy schema resolution failures', () => {
+        const lazySchema = z.lazy(() => {
+          throw new Error('Lazy schema failed');
+        });
+
+        const result = convertZodToJsonSchema(lazySchema);
+
+        expect(result.schema).toEqual({
+          'x-zod': {
+            unmapped: true,
+            reason: 'Failed to resolve lazy schema',
+          },
+        });
+        expect(result.hasUnmappedNodes).toBe(true);
+        expect(result.warnings).toContain('Failed to resolve lazy schema');
+      });
+
+      it('should handle maximum depth exceeded', () => {
+        // Create a deeply nested schema
+        let deepSchema: z.ZodTypeAny = z.string();
+        for (let i = 0; i < 15; i++) {
+          deepSchema = z.object({ nested: deepSchema });
+        }
+
+        const result = convertZodToJsonSchema(deepSchema, { maxDepth: 5 });
+
+        expect(result.warnings.some(w => w.includes('Maximum depth'))).toBe(
+          true
+        );
+      });
+
+      it('should handle circular references', () => {
+        interface Node {
+          value: string;
+          child?: Node;
+        }
+
+        const nodeSchema: z.ZodType<Node> = z.lazy(() =>
+          z.object({
+            value: z.string(),
+            child: nodeSchema.optional(),
+          })
+        );
+
+        const result = convertZodToJsonSchema(nodeSchema);
+
+        // Should complete without infinite recursion
+        expect(result.schema.type).toBe('object');
+        expect(result.schema.properties?.value).toEqual({ type: 'string' });
+      });
+    });
+
+    describe('x-zod annotations for refinements', () => {
+      it('should annotate custom refinements', () => {
+        const schema = z
+          .string()
+          .refine(val => val.length > 5, 'Must be longer than 5 characters');
+
+        const result = convertZodToJsonSchema(schema);
+
+        expect(result.schema).toEqual({
+          type: 'string',
+          'x-zod': {
+            refinements: ['custom-refinement'],
+          },
+        });
+      });
+
+      it('should annotate transform operations', () => {
+        const schema = z.string().transform(val => val.toUpperCase());
+
+        const result = convertZodToJsonSchema(schema);
+
+        expect(result.schema).toEqual({
+          type: 'string',
+          'x-zod': {
+            refinements: ['transform'],
+          },
+        });
+      });
+
+      it('should annotate preprocess operations', () => {
+        const schema = z.preprocess(val => String(val), z.string());
+
+        const result = convertZodToJsonSchema(schema);
+
+        expect(result.schema).toEqual({
+          type: 'string',
+          'x-zod': {
+            refinements: ['preprocess'],
+          },
+        });
+      });
+
+      it('should annotate multiple refinements', () => {
+        const schema = z
+          .string()
+          .refine(val => val.length > 3, 'Too short')
+          .transform(val => val.trim())
+          .refine(val => !val.includes(' '), 'No spaces allowed');
+
+        const result = convertZodToJsonSchema(schema);
+
+        expect(result.schema).toEqual({
+          type: 'string',
+          'x-zod': {
+            refinements: [
+              'custom-refinement',
+              'transform',
+              'custom-refinement',
+            ],
+          },
+        });
+      });
+    });
+
+    describe('complex scenarios with mixed success and failure', () => {
+      it('should handle mixed success and failure in objects', () => {
+        const validSchema = z.string();
+        const unsupportedSchema = z.string();
+
+        // Make one schema unsupported
+        Object.defineProperty(unsupportedSchema._def, 'typeName', {
+          value: 'ZodUnsupported',
+          writable: false,
+        });
+
+        const objectSchema = z.object({
+          validProp: validSchema,
+          unsupportedProp: unsupportedSchema,
+          anotherValidProp: z.number(),
+        });
+
+        const result = convertZodToJsonSchema(objectSchema);
+
+        expect(result.schema.type).toBe('object');
+        expect(result.schema.properties?.validProp).toEqual({ type: 'string' });
+        expect(result.schema.properties?.anotherValidProp).toEqual({
+          type: 'number',
+        });
+        expect(result.schema.properties?.unsupportedProp).toEqual({
+          'x-zod': {
+            unmapped: true,
+            reason: 'Unsupported type: ZodUnsupported',
+          },
+        });
+        expect(result.hasUnmappedNodes).toBe(true);
+        expect(result.schema.required).toEqual([
+          'validProp',
+          'unsupportedProp',
+          'anotherValidProp',
+        ]);
+      });
+
+      it('should accumulate warnings from multiple failures', () => {
+        const schema1 = z.string();
+        const schema2 = z.number();
+
+        Object.defineProperty(schema1._def, 'typeName', {
+          value: 'ZodUnsupported1',
+          writable: false,
+        });
+
+        Object.defineProperty(schema2._def, 'typeName', {
+          value: 'ZodUnsupported2',
+          writable: false,
+        });
+
+        const unionSchema = z.union([schema1, schema2]);
+
+        const result = convertZodToJsonSchema(unionSchema);
+
+        expect(result.warnings).toHaveLength(2);
+        expect(result.warnings.some(w => w.includes('ZodUnsupported1'))).toBe(
+          true
+        );
+        expect(result.warnings.some(w => w.includes('ZodUnsupported2'))).toBe(
+          true
+        );
+        expect(result.hasUnmappedNodes).toBe(true);
+      });
+    });
+  });
 });
