+++ packages/express/src/__tests__/integration.test.ts
@@ -1,3 +1,4 @@
+import { describe, it, expect, beforeEach, afterEach } from 'vitest';
 import express from 'express';
 import request from 'supertest';
 import { z } from 'zod';
@@ -364,6 +365,240 @@ describe('Express Studio Integration', () => {
       );
       expect(createProcedure).toBeDefined();
       expect(createProcedure.type).toBe('mutation');
+    });
+  });
+
+  describe('Security Integration', () => {
+    describe('Production Security', () => {
+      it('should be disabled by default in production', async () => {
+        process.env.NODE_ENV = 'production';
+
+        const app = express();
+        app.use(express.json());
+
+        const options: ExpressStudioOptions = {
+          router: mockRouter,
+        };
+
+        app.use(studioExpress(options));
+
+        // All studio routes should return 404
+        await request(app).get('/trpc-studio').expect(404);
+        await request(app).get('/__trpc-studio__/introspection').expect(404);
+      });
+
+      it('should require TRPC_STUDIO_ENABLED=true in production', async () => {
+        process.env.NODE_ENV = 'production';
+        process.env.TRPC_STUDIO_TOKEN = 'test-token';
+
+        const app = express();
+        app.use(express.json());
+
+        const options: ExpressStudioOptions = {
+          router: mockRouter,
+          enabled: true, // This should be ignored in production
+        };
+
+        app.use(studioExpress(options));
+
+        // Should still be disabled despite options.enabled = true
+        await request(app).get('/trpc-studio').expect(404);
+        await request(app).get('/__trpc-studio__/introspection').expect(404);
+      });
+
+      it('should work with both env flag and token in production', async () => {
+        process.env.NODE_ENV = 'production';
+        process.env.TRPC_STUDIO_ENABLED = 'true';
+        process.env.TRPC_STUDIO_TOKEN = 'secure-token-123';
+
+        const app = express();
+        app.use(express.json());
+
+        const options: ExpressStudioOptions = {
+          router: mockRouter,
+        };
+
+        app.use(studioExpress(options));
+
+        // Should work with valid token
+        await request(app)
+          .get('/__trpc-studio__/introspection')
+          .set('Authorization', 'Bearer secure-token-123')
+          .expect(200);
+
+        await request(app)
+          .get('/trpc-studio')
+          .set('Authorization', 'Bearer secure-token-123')
+          .expect(200);
+
+        // Should fail without token
+        await request(app).get('/__trpc-studio__/introspection').expect(403);
+        await request(app).get('/trpc-studio').expect(403);
+
+        // Should fail with wrong token
+        await request(app)
+          .get('/__trpc-studio__/introspection')
+          .set('Authorization', 'Bearer wrong-token')
+          .expect(403);
+      });
+    });
+
+    describe('Token Validation', () => {
+      beforeEach(() => {
+        process.env.NODE_ENV = 'production';
+        process.env.TRPC_STUDIO_ENABLED = 'true';
+        process.env.TRPC_STUDIO_TOKEN = 'test-token-123';
+      });
+
+      it('should accept Authorization Bearer token', async () => {
+        const app = express();
+        app.use(express.json());
+
+        const options: ExpressStudioOptions = {
+          router: mockRouter,
+        };
+
+        app.use(studioExpress(options));
+
+        await request(app)
+          .get('/__trpc-studio__/introspection')
+          .set('Authorization', 'Bearer test-token-123')
+          .expect(200);
+      });
+
+      it('should accept x-trpc-studio-token header', async () => {
+        const app = express();
+        app.use(express.json());
+
+        const options: ExpressStudioOptions = {
+          router: mockRouter,
+        };
+
+        app.use(studioExpress(options));
+
+        await request(app)
+          .get('/__trpc-studio__/introspection')
+          .set('x-trpc-studio-token', 'test-token-123')
+          .expect(200);
+      });
+
+      it('should use custom token extraction function', async () => {
+        const app = express();
+        app.use(express.json());
+
+        const options: ExpressStudioOptions = {
+          router: mockRouter,
+          getToken: (req: any) => {
+            // Extract token from custom header
+            return req.headers['x-custom-token'] || null;
+          },
+        };
+
+        app.use(studioExpress(options));
+
+        await request(app)
+          .get('/__trpc-studio__/introspection')
+          .set('x-custom-token', 'test-token-123')
+          .expect(200);
+
+        // Standard headers should not work with custom extraction
+        await request(app)
+          .get('/__trpc-studio__/introspection')
+          .set('Authorization', 'Bearer test-token-123')
+          .expect(403);
+      });
+
+      it('should handle malformed Authorization header', async () => {
+        const app = express();
+        app.use(express.json());
+
+        const options: ExpressStudioOptions = {
+          router: mockRouter,
+        };
+
+        app.use(studioExpress(options));
+
+        // Malformed Authorization header
+        await request(app)
+          .get('/__trpc-studio__/introspection')
+          .set('Authorization', 'InvalidFormat test-token-123')
+          .expect(403);
+
+        await request(app)
+          .get('/__trpc-studio__/introspection')
+          .set('Authorization', 'Bearer')
+          .expect(403);
+      });
+    });
+
+    describe('Development Security', () => {
+      beforeEach(() => {
+        process.env.NODE_ENV = 'development';
+        delete process.env.TRPC_STUDIO_ENABLED;
+        delete process.env.TRPC_STUDIO_TOKEN;
+      });
+
+      it('should be enabled by default in development', async () => {
+        const app = express();
+        app.use(express.json());
+
+        const options: ExpressStudioOptions = {
+          router: mockRouter,
+        };
+
+        app.use(studioExpress(options));
+
+        // Should work without token in development
+        await request(app).get('/__trpc-studio__/introspection').expect(200);
+        await request(app).get('/trpc-studio').expect(200);
+      });
+
+      it('should respect explicit disabled option in development', async () => {
+        const app = express();
+        app.use(express.json());
+
+        const options: ExpressStudioOptions = {
+          router: mockRouter,
+          enabled: false,
+        };
+
+        app.use(studioExpress(options));
+
+        // Should be disabled when explicitly set
+        await request(app).get('/__trpc-studio__/introspection').expect(404);
+        await request(app).get('/trpc-studio').expect(404);
+      });
+    });
+
+    describe('HTTP Method Security', () => {
+      beforeEach(() => {
+        process.env.NODE_ENV = 'development';
+        delete process.env.TRPC_STUDIO_ENABLED;
+        delete process.env.TRPC_STUDIO_TOKEN;
+      });
+
+      it('should only allow GET requests to introspection endpoint', async () => {
+        const app = express();
+        app.use(express.json());
+
+        const options: ExpressStudioOptions = {
+          router: mockRouter,
+        };
+
+        app.use(studioExpress(options));
+
+        // GET should work
+        await request(app).get('/__trpc-studio__/introspection').expect(200);
+
+        // Other methods should return 405
+        const response = await request(app)
+          .post('/__trpc-studio__/introspection')
+          .expect(405);
+        expect(response.headers.allow).toBe('GET');
+
+        await request(app).put('/__trpc-studio__/introspection').expect(405);
+        await request(app).delete('/__trpc-studio__/introspection').expect(405);
+      });
     });
   });
 });
