--- packages/ui/src/__tests__/ci-integration.test.ts
@@ -1,474 +1,0 @@
-/**
- * CI integration and coverage tests
- */
-
-import { describe, it, expect, vi } from 'vitest';
-import { readFileSync, existsSync } from 'fs';
-import { resolve } from 'path';
-
-describe('CI Integration', () => {
-  describe('Coverage Configuration', () => {
-    it('should have coverage configuration in vitest config', () => {
-      const vitestConfigPath = resolve(__dirname, '../../../vitest.config.ts');
-
-      if (!existsSync(vitestConfigPath)) {
-        console.warn('vitest.config.ts not found, skipping coverage test');
-        return;
-      }
-
-      const config = readFileSync(vitestConfigPath, 'utf-8');
-
-      // Should have coverage configuration
-      expect(config).toContain('coverage');
-      expect(config).toContain('provider');
-      expect(config).toContain('reporter');
-
-      // Should exclude test files from coverage
-      expect(config).toContain('exclude');
-      expect(config).toMatch(/\*\*\/\*\.test\./);
-      expect(config).toMatch(/\*\*\/__tests__/);
-    });
-
-    it('should meet coverage thresholds', async () => {
-      // Mock coverage data that would be generated by vitest
-      const mockCoverageData = {
-        lines: { pct: 85 },
-        statements: { pct: 87 },
-        functions: { pct: 82 },
-        branches: { pct: 78 },
-      };
-
-      // Coverage thresholds
-      const thresholds = {
-        lines: 80,
-        statements: 80,
-        functions: 75,
-        branches: 70,
-      };
-
-      expect(mockCoverageData.lines.pct).toBeGreaterThanOrEqual(
-        thresholds.lines
-      );
-      expect(mockCoverageData.statements.pct).toBeGreaterThanOrEqual(
-        thresholds.statements
-      );
-      expect(mockCoverageData.functions.pct).toBeGreaterThanOrEqual(
-        thresholds.functions
-      );
-      expect(mockCoverageData.branches.pct).toBeGreaterThanOrEqual(
-        thresholds.branches
-      );
-    });
-
-    it('should generate coverage reports in multiple formats', () => {
-      const coverageDir = resolve(__dirname, '../../../coverage');
-
-      // Expected coverage report files
-      const expectedReports = [
-        'lcov.info',
-        'coverage-final.json',
-        'index.html',
-      ];
-
-      for (const report of expectedReports) {
-        const reportPath = resolve(coverageDir, report);
-
-        if (existsSync(reportPath)) {
-          console.log(`Coverage report found: ${report}`);
-
-          if (report === 'lcov.info') {
-            const lcovContent = readFileSync(reportPath, 'utf-8');
-            expect(lcovContent).toContain('TN:'); // Test name
-            expect(lcovContent).toContain('SF:'); // Source file
-            expect(lcovContent).toContain('LF:'); // Lines found
-            expect(lcovContent).toContain('LH:'); // Lines hit
-          }
-        } else {
-          console.warn(
-            `Coverage report not found: ${report} (may not be generated yet)`
-          );
-        }
-      }
-    });
-  });
-
-  describe('Test Organization', () => {
-    it('should have tests organized by package', () => {
-      const packages = ['core', 'ui', 'next', 'express'];
-
-      for (const pkg of packages) {
-        const testDir = resolve(__dirname, `../../../${pkg}/src/__tests__`);
-
-        if (existsSync(testDir)) {
-          console.log(`Test directory found for package: ${pkg}`);
-        } else {
-          console.warn(`Test directory not found for package: ${pkg}`);
-        }
-      }
-    });
-
-    it('should have consistent test naming patterns', () => {
-      const testPatterns = [
-        /\.test\.(ts|tsx|js|jsx)$/,
-        /\.spec\.(ts|tsx|js|jsx)$/,
-      ];
-
-      // This would be expanded to actually scan test files
-      const mockTestFiles = [
-        'component.test.tsx',
-        'utils.test.ts',
-        'integration.spec.ts',
-        'e2e.test.ts',
-      ];
-
-      for (const testFile of mockTestFiles) {
-        const matchesPattern = testPatterns.some(pattern =>
-          pattern.test(testFile)
-        );
-        expect(matchesPattern).toBe(true);
-      }
-    });
-
-    it('should have test setup files configured', () => {
-      const setupFiles = [
-        resolve(__dirname, '../../../src/test-setup.ts'),
-        resolve(__dirname, '../../../../express/src/test-setup.ts'),
-      ];
-
-      for (const setupFile of setupFiles) {
-        if (existsSync(setupFile)) {
-          const content = readFileSync(setupFile, 'utf-8');
-
-          // Should have basic test setup
-          expect(content).toMatch(/(beforeEach|afterEach|vi\.)/);
-
-          console.log(`Test setup file found: ${setupFile}`);
-        }
-      }
-    });
-  });
-
-  describe('Test Categories', () => {
-    it('should have unit tests for core functionality', () => {
-      const coreTestCategories = [
-        'introspection',
-        'schema conversion',
-        'security',
-        'token validation',
-      ];
-
-      // Mock test results for each category
-      const mockTestResults = {
-        introspection: { passed: 25, failed: 0 },
-        'schema conversion': { passed: 30, failed: 0 },
-        security: { passed: 15, failed: 0 },
-        'token validation': { passed: 12, failed: 0 },
-      };
-
-      for (const category of coreTestCategories) {
-        const results =
-          mockTestResults[category as keyof typeof mockTestResults];
-        expect(results.passed).toBeGreaterThan(0);
-        expect(results.failed).toBe(0);
-      }
-    });
-
-    it('should have integration tests for adapters', () => {
-      const adapterTestCategories = [
-        'next.js app router',
-        'next.js pages router',
-        'express middleware',
-        'cors handling',
-        'security integration',
-      ];
-
-      // Mock test results for each category
-      const mockTestResults = {
-        'next.js app router': { passed: 20, failed: 0 },
-        'next.js pages router': { passed: 18, failed: 0 },
-        'express middleware': { passed: 22, failed: 0 },
-        'cors handling': { passed: 15, failed: 0 },
-        'security integration': { passed: 25, failed: 0 },
-      };
-
-      for (const category of adapterTestCategories) {
-        const results =
-          mockTestResults[category as keyof typeof mockTestResults];
-        expect(results.passed).toBeGreaterThan(0);
-        expect(results.failed).toBe(0);
-      }
-    });
-
-    it('should have e2e tests for critical user flows', () => {
-      const e2eTestCategories = [
-        'studio loading',
-        'procedure interaction',
-        'form submission',
-        'environment switching',
-        'accessibility',
-      ];
-
-      // Mock test results for each category
-      const mockTestResults = {
-        'studio loading': { passed: 5, failed: 0 },
-        'procedure interaction': { passed: 8, failed: 0 },
-        'form submission': { passed: 6, failed: 0 },
-        'environment switching': { passed: 4, failed: 0 },
-        accessibility: { passed: 3, failed: 0 },
-      };
-
-      for (const category of e2eTestCategories) {
-        const results =
-          mockTestResults[category as keyof typeof mockTestResults];
-        expect(results.passed).toBeGreaterThan(0);
-        expect(results.failed).toBe(0);
-      }
-    });
-  });
-
-  describe('Performance Tests', () => {
-    it('should have bundle size tests', () => {
-      // This test file itself is a bundle size test
-      expect(true).toBe(true);
-    });
-
-    it('should have performance benchmarks', () => {
-      // Mock performance test results
-      const performanceMetrics = {
-        'introspection build time': 50, // ms
-        'schema conversion time': 25, // ms
-        'ui render time': 100, // ms
-        'form generation time': 75, // ms
-      };
-
-      const performanceBudgets = {
-        'introspection build time': 100,
-        'schema conversion time': 50,
-        'ui render time': 200,
-        'form generation time': 150,
-      };
-
-      for (const [metric, value] of Object.entries(performanceMetrics)) {
-        const budget =
-          performanceBudgets[metric as keyof typeof performanceBudgets];
-        expect(value).toBeLessThanOrEqual(budget);
-      }
-    });
-
-    it('should test memory usage', () => {
-      // Mock memory usage test
-      const memoryUsage = {
-        'large router introspection': 50, // MB
-        'complex form rendering': 25, // MB
-        'response panel with large data': 30, // MB
-      };
-
-      const memoryBudgets = {
-        'large router introspection': 100,
-        'complex form rendering': 50,
-        'response panel with large data': 75,
-      };
-
-      for (const [test, usage] of Object.entries(memoryUsage)) {
-        const budget = memoryBudgets[test as keyof typeof memoryBudgets];
-        expect(usage).toBeLessThanOrEqual(budget);
-      }
-    });
-  });
-
-  describe('Test Quality', () => {
-    it('should have descriptive test names', () => {
-      const goodTestNames = [
-        'should render documentation view with router tree',
-        'should validate token from Authorization header',
-        'should return 404 when studio is disabled in production',
-        'should generate cURL command with custom headers',
-      ];
-
-      const badTestNames = [
-        'test1',
-        'it works',
-        'check function',
-        'basic test',
-      ];
-
-      for (const goodName of goodTestNames) {
-        // Good test names should be descriptive
-        expect(goodName.length).toBeGreaterThan(20);
-        expect(goodName).toMatch(/should|when|if|given/);
-      }
-
-      for (const badName of badTestNames) {
-        // These would fail our naming standards
-        expect(badName.length).toBeLessThan(50); // Just to show they're short
-      }
-    });
-
-    it('should have proper test isolation', () => {
-      // Mock test isolation checks
-      const testIsolationChecks = {
-        'beforeEach cleanup': true,
-        'afterEach cleanup': true,
-        'no shared state': true,
-        'mocks reset': true,
-        'environment reset': true,
-      };
-
-      for (const [check, passes] of Object.entries(testIsolationChecks)) {
-        expect(passes).toBe(true);
-      }
-    });
-
-    it('should have appropriate test timeouts', () => {
-      const testTimeouts = {
-        'unit tests': 5000, // 5s
-        'integration tests': 10000, // 10s
-        'e2e tests': 30000, // 30s
-      };
-
-      const timeoutLimits = {
-        'unit tests': 10000,
-        'integration tests': 20000,
-        'e2e tests': 60000,
-      };
-
-      for (const [testType, timeout] of Object.entries(testTimeouts)) {
-        const limit = timeoutLimits[testType as keyof typeof timeoutLimits];
-        expect(timeout).toBeLessThanOrEqual(limit);
-      }
-    });
-  });
-
-  describe('CI Pipeline Integration', () => {
-    it('should have GitHub Actions workflow', () => {
-      const workflowPath = resolve(
-        __dirname,
-        '../../../../../.github/workflows/ci.yml'
-      );
-
-      if (existsSync(workflowPath)) {
-        const workflow = readFileSync(workflowPath, 'utf-8');
-
-        // Should test on multiple Node versions
-        expect(workflow).toMatch(/node.*20|22/);
-
-        // Should have required jobs
-        expect(workflow).toContain('typecheck');
-        expect(workflow).toContain('lint');
-        expect(workflow).toContain('test');
-        expect(workflow).toContain('build');
-
-        // Should use pnpm
-        expect(workflow).toContain('pnpm');
-
-        // Should have caching
-        expect(workflow).toMatch(/cache|restore/);
-
-        console.log('GitHub Actions workflow found and validated');
-      } else {
-        console.warn('GitHub Actions workflow not found');
-      }
-    });
-
-    it('should fail CI on test failures', () => {
-      // Mock CI behavior
-      const ciResults = {
-        typecheck: 'passed',
-        lint: 'passed',
-        test: 'passed',
-        build: 'passed',
-        'bundle-size': 'passed',
-      };
-
-      for (const [step, result] of Object.entries(ciResults)) {
-        expect(result).toBe('passed');
-      }
-    });
-
-    it('should report coverage to external services', () => {
-      // Mock coverage reporting
-      const coverageReporting = {
-        'codecov upload': true,
-        'coverage comment': true,
-        'coverage badge': true,
-      };
-
-      for (const [service, enabled] of Object.entries(coverageReporting)) {
-        if (enabled) {
-          console.log(`Coverage reporting enabled for: ${service}`);
-        }
-      }
-    });
-  });
-
-  describe('Test Data Management', () => {
-    it('should have consistent mock data', () => {
-      // Mock data should be consistent across tests
-      const mockData = {
-        router: {
-          name: 'root',
-          procedures: [
-            { name: 'hello', type: 'query' },
-            { name: 'createUser', type: 'mutation' },
-          ],
-        },
-        environment: {
-          id: 'test',
-          name: 'Test Environment',
-          baseUrl: 'https://api.example.com',
-        },
-      };
-
-      expect(mockData.router.name).toBe('root');
-      expect(mockData.router.procedures).toHaveLength(2);
-      expect(mockData.environment.baseUrl).toMatch(/^https?:\/\//);
-    });
-
-    it('should have test fixtures for complex scenarios', () => {
-      // Mock test fixtures
-      const fixtures = {
-        'large router': { procedures: 100, routers: 10 },
-        'complex schema': { properties: 20, nested: 5 },
-        'superjson response': { types: ['Date', 'BigInt', 'Set'] },
-      };
-
-      for (const [fixture, config] of Object.entries(fixtures)) {
-        expect(config).toBeDefined();
-        console.log(`Test fixture available: ${fixture}`);
-      }
-    });
-  });
-
-  describe('Test Environment', () => {
-    it('should have proper test environment setup', () => {
-      // Check test environment configuration
-      expect(process.env.NODE_ENV).toBe('test');
-
-      // Should have test-specific configurations
-      const testConfig = {
-        jsdom: true,
-        mocks: true,
-        coverage: true,
-        parallel: true,
-      };
-
-      for (const [config, enabled] of Object.entries(testConfig)) {
-        expect(enabled).toBe(true);
-      }
-    });
-
-    it('should isolate tests from external dependencies', () => {
-      // Mock external dependencies
-      const externalDeps = {
-        fetch: 'mocked',
-        localStorage: 'mocked',
-        clipboard: 'mocked',
-        ResizeObserver: 'mocked',
-      };
-
-      for (const [dep, status] of Object.entries(externalDeps)) {
-        expect(status).toBe('mocked');
-      }
-    });
-  });
-});
